config.py
import os
from dotenv import load_dotenv

load_dotenv("api.env")  

MISTRAL_API_KEY = os.getenv("MISTRAL_API_KEY")
if not MISTRAL_API_KEY:
    raise ValueError("MISTRAL_API_KEY environment variable not set!")
