requirements.txt
streamlit
requests
python-dotenv
camel-ai
mistral